Added some ajax, and tweaks in templates

+++ src/main/java/pl/codespring/jTask/task/TaskService.java
@@ -28,6 +28,8 @@ public class TaskService {
     }
 
     public void deleteTask(int id){
-        taskRepository.deleteById(id);
+        if(taskRepository.existsById(id)) {
+            taskRepository.deleteById(id);
+        }
     }
 }
